+++ src/main/java/com/luan/getlib/models/Book.java
@@ -1,16 +1,39 @@
 package com.luan.getlib.models;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
 /**
  * @since v0.2.0
  * @author luanpozzobon
  */
+
+@Entity
+@Table(name = "books")
 public class Book {
+    
+    @Id
+    @Column(name = "book_id")
     private int id;
+    
+    @Column(name = "title")
     private String title;
+    
+    @Column(name = "genre")
     private String genre;
+    
+    @Column(name = "amount")
     private int amount;
+    
+    @Column(name = "value")
     private double value;
+    
+    @Column(name = "parentalRating")
     private int parentalRating;
+    
+    public Book(){ }
     
     public Book(String title, String genre, int amount, double value, int parentalRating) {
         this.title = title;
@@ -18,20 +41,6 @@ public class Book {
         this.amount = amount;
         this.value = value;
         this.parentalRating = parentalRating;
-    }
-    
-    public Book(int id, String title, String genre, int amount, double value, int parentalRating) {
-        this.id = id;
-        this.title = title;
-        this.genre = genre;
-        this.amount = amount;
-        this.value = value;
-        this.parentalRating = parentalRating;
-    }
-    
-    public Book(int id, String title){
-        this.id = id;
-        this.title = title;
     }
 
     public int getId() {